requirements.txt
transformers
tensorboard
sentencepiece
pandas
scikit-learn
langchain
python-dotenv
spacy